EU variants: Update patch for CPE210
To avoid a patch to a patch, here manually inserted the extra lines to support EU|US|CN flavours of the CP210 model. Admittedly I do not really know what I am doing here as modifying the files directly may interfere with the semantics of the interplay between git and quilt within gluon.

This change was proposed and reported tested here https://forum.openwrt.org/viewtopic.php?id=64748 . Not unexpectedly, it also motivates (untested, though) the introduction of analogous models for the CPE220 and CPE5[12]0.